Make a better caching mechanism.
The problem with the old caching was that the entire dataset would be
stored, which means that every time the dataset is loaded from disk,
that were saved at the time would also be loaded. In other words,
changes in the IAMDataset class would not have an effect if the dataset
was loaded from cache.

The way this is fixed now is by only caching the label-index mapping
that the l2l MetaDataset requires, which takes up the bulk of the
initialization time.